off screen anims need to die
Former-commit-id: a6e4bc346ccb8413495b38a83b258e7612541349

+++ src/my_game.h
@@ -223,7 +223,7 @@ public:
     uint32_t           skillbox_highlight_slot {};
     uint32_t           tick_completed {1};
     uint32_t           tick_current {1};
-    uint32_t           frame_count {0};
+    uint16_t           frame_count {0};     // Used to know if things have been displayed
 
     /////////////////////////////////////////////////////////////////////////
     // not worth saving
